Unify `ProtobufJsonReader` and `ProtobufJsonWriter` implementations into single classes.
Motivation:

- Simplify the codebase by eliminating separate `Impl` classes that provide no additional value.
- Improve clarity by making `ProtobufJsonReader` and `ProtobufJsonWriter` self-contained.

Changes:

- Merged `ProtobufJsonReaderImpl` into `ProtobufJsonReader`.
- Merged `ProtobufJsonWriterImpl` into `ProtobufJsonWriter`.
- Updated constructors to handle `JsonWireFormat` directly within unified classes.
- Removed redundant `Impl` classes and adjusted usage across the codebase.
